build `nimsuggest_testing` for local testing (#900)
<!--- The Pull Request (=PR) message is what will get automatically used
as
the commit message when the PR is merged. Make sure that no line is
longer
than 72 characters -->

## Summary

Make local nimsuggest testing easier by having the nimsuggest `tester`
build the `nimsuggest_testing` binary as part of the run.

## Details

Previous `nimsuggest_testing` is built in CI environment which meant it
is not available when running tests locally. Now if `tester` can't find
the `nimsuggest_testing` binary, it's built before running any tests.

To test locally, as per the module comments:
> When debugging, to run a single test, use for e.g.:
> `nim r nimsuggest/tester.nim nimsuggest/tests/tsug_accquote.nim`

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>