internal: make NDI mappings independent from `PSym`
Summary
=======

Change how symbol-to-name mappings are stored in an `NdiFile` such
that its not dependent on getting a `PSym` as the input. This is a
preparation for moving away from using `PSym` in the code generators.

Details
=======

Only the source position (`TLineInfo`) and the user-provided name
(`PIdent`) were previously taken from the `PSym`. Both are now stored in
the mapping entry directly, making it possible to register a mapping
when not having access to `PSym`.

A new and exported, but currently unused, `writeMangledName` overload is
added, which allows for passing the `TLineInfo` and `PIdent` directly.