vm: convert `gorge` and `staticExec` to vmops (#878)
<!--- The Pull Request (=PR) message is what will get automatically used
as
the commit message when the PR is merged. Make sure that no line is
longer
than 72 characters -->

## Summary

Convert `gorge` and `staticExec` in the `system` module from magics to
VM callback operations.

## Details

Removal of the magic requires creating a conditional compilation step
based on the `nimskullReworkStaticExec` conditional symbol (see:
`condsym`). Two definitions of `gorge` and `staticExec` now exist, one
with the magic and one without. Without this, bootstrapping would no
longer work.

Also marked `gorge` as deprecated, `staticExec` is a far better name,
and noted that `staticExec` will have its return type revised to return
the exit code in addition to the output. The reason for the deprecation
is that `gorge` is just not as clear as `staticExec`, same for `slurp`.

This change enables future removal of the `mStaticExec` magic enum, and
more importantly the `opcGorge` VM opcode in the near future.