internal: make NDI mappings independent from `PSym` (#906)
## Summary

Change how symbol-to-name mappings are stored in an `NdiFile` such
that it's not dependent on getting a `PSym` as the input. This is a
preparation for moving away from using `PSym` in the code generators.

## Details

Only the source position (`TLineInfo`) and the user-provided name
(`PIdent`) were previously taken from the `PSym`. Both are now stored in
the mapping entry directly, making it possible to register a mapping
when not having access to `PSym`.

A new and exported, but currently unused, `writeMangledName` overload is
added, which allows for passing the `TLineInfo` and `PIdent` directly.

To aid with debugging, an additional ID value is stored in the mapping
entry (but it's not written to the `.ndi` file). This value is meant for
tracing the entry back to originating symbol.