remove disabled code in `modulepaths.nim` (#911)
<!--- The Pull Request (=PR) message is what will get automatically used
as
the commit message when the PR is merged. Make sure that no line is
longer
than 72 characters -->

## Summary

Remove disabled code related to `nimble` in `modulepaths`, it was left
over from a long ago (6+ years) disabled import feature related to
nimble. Since we'll have a new package manager anyways, not bothering
to investigate it any further.

## Details
* Remove constants `considerParentDirs`, `considerNimbleDirs`
* Remove procedures `findInNimbleDir`,
`resolveDollar`,`scriptableImport`,
 `lookupPackage`
* Remove `lookupPackage` disabled usage in `getModuleName`