tool: nimsuggest remove `terse` flag (#898)
<!--- The Pull Request (=PR) message is what will get automatically used
as
the commit message when the PR is merged. Make sure that no line is
longer
than 72 characters -->

## Summary

Remove `terse` flag from `nimsuggest`. It's not used by any of the
existing extensions, nor is it tested.

## Details

The option itself is rather arbitrary in what it filters and the design
is confusing in terms of determining what should and shouldn't qualify.