restore incremental procedure code generation
When incrementally generating the target-language code for a procedure,
the respective code generator's `OrdinalSeq` storing the state of the
locals needs to be synchronized with the data from `Body`.

The new `genPartial` procedure added to both the C and JS code
generators takes care of this; the VM code generator doesn't use
incremental code generation for procedures and thus doesn't need a
`genPartial` procedure.