jsgen: better code generation for labeled blocks
Summary
=======

Instead of `Label: do {} while(false)`, emit a `Label: {}` for labeled
block statements. This shrinks the size of the generated code.

Details
=======

ECMAScript 5 (that is, the version targeted by the JS code generator)
allows attaching labels to blocks (`{}`), which is less code than using
a do-while statement.

In theory, the wrapping curlies could be elided if the content of the
`cnkBlockStmt`/procedure is a statement (and thus can be labeled
directly), but this would presently introduce a not-immediately-obvious
dependency between the emitting code (e.g., `genBlock`) and code
generation for the statement (e.g., `genRepeatStmt`), so in order to not
hinder evolution of `jsgen`, this optimization is not applied for now.