remove `ConfigRef.projectMainIdx2` (#922)
## Summary

Remove  `ConfigRef.projectMainIdx2` , this has no user facing impact,
it's
just poor design/implementation.

## Details

Remove  `ConfigRef.projectMainIdx2` , which was used to handle cases
where
a project was being recompiled during  `suggest`  but with a dirty
project
module, or a module outside of the scope of the project. It's
ultimately
unnecessary as module, not project, recompilation should be sufficient.

All usages were in the `modules.compileProject` and have been replaced
with `ConfigRef.projectMainIdx`.