## Summary
Unmatched `immediate` `macro`s and `template`s within a generic context
no longer result in a compiler error.

## Details

Immediate macros and templates are not dispatched via `sigmatch` and
instead `semgnrc` directly calls `evaltempl.evalTemplate`, in turn
`evalTemplateArgs` is called which can produce an error if an incorrect
number of arguments are provided. `evalTemplateArgs` is also used by
`semMacroExpr`, so similar issues would also exist for macros.

Now the error from `evalTemplateArgs` is immediately returned by
`evalTemplate` and treated as a mismatch in `semgnrc.semGenericStmt`.
Doing so prevents a compiler crash, attempt to access the sons of an error
node. A test has been added as part of this change, which was originally
from: https://github.com/zevv/npeg/issues/68#issuecomment-1730098232